requirements.txt
requests
beautifulsoup4
lxml
pandas
pyarrow
google-cloud-storage
google-cloud-bigquery
jupyter
matplotlib
pycountry
